ref:  ts, prettier, esLint, and restructuraction project
