+++ src/constants/genericInterfaces.ts
@@ -1,0 +1,3 @@
+export interface GenericProps {
+  id: string;
+}
