+++ src/fragments/Footer.tsx
@@ -1,12 +1,13 @@
+import React from 'react';
+
 import '../styles/footer.css';
 import texts from '../utils/texts';
+
 const FooterFragment = () => {
   return (
-    <>
-      <div className="footer-container">
-        <p>{texts.footer_text}</p>
-      </div>
-    </>
+    <div className="footer-container">
+      <p>{texts.footer_text}</p>
+    </div>
   );
 };
