+++ src/main.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import ReactDOM from 'react-dom/client';
+
+import App from './pages/app';
+import './theme/utils.scss';
+
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
