+++ src/pages/sections/Profile/Profile.scss
@@ -1,0 +1,73 @@
+@use '../../../theme/styles.scss' as *;
+@use '../../../theme/colors.scss' as *;
+
+.profile {
+  padding-top: 50px;
+  margin: 0;
+  width: 100%;
+  padding-bottom: 10px;
+  margin-top: 50px;
+  min-width: 400px;
+
+  .info {
+    padding: 0;
+    margin: 0;
+    width: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-direction: row;
+    gap: 20%;
+    .photo {
+      position: relative;
+      height: 250px;
+      width: 250px;
+      border-radius: 100%;
+      background: $gradient-bottom-topRight;
+      box-shadow: 0 4px 6px rgba(0, 0, 0, 0.19);
+      img {
+        width: 100%;
+        height: 100%;
+        border-radius: inherit;
+      }
+    }
+
+    .me {
+      width: 20%;
+      p:nth-child(1) {
+        background: $gradient-left-right;
+        -webkit-background-clip: text;
+        background-clip: text;
+        color: transparent;
+      }
+      p:nth-child(1) {
+        font-size: 45px;
+        font-weight: bold;
+        line-height: 1;
+      }
+      p:nth-child(2) {
+        font-size: 20px;
+        font-weight: 500;
+        line-height: 1;
+      }
+
+      p:nth-child(3) {
+        font-size: 15px;
+        line-height: 1.5;
+      }
+      span {
+        color: $primary-color;
+      }
+    }
+  }
+
+  @media (max-width: 750px) {
+    .info {
+      flex-direction: column-reverse;
+      .me {
+        width: 70%;
+        text-align: center;
+      }
+    }
+  }
+}
