--- src/styles/profile.css
@@ -1,109 +1,0 @@
-.profile-container {
-  padding-top: 50px;
-  margin: 0;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding-bottom: 10px;
-  margin-top: 50px;
-  min-width: 400px;
-}
-
-.profile-container--info {
-  padding: 0;
-  margin: 0;
-  width: 100%;
-
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  flex-direction: row;
-}
-.profile-container--info-photo {
-  position: relative;
-  height: 250px;
-}
-.profile-container--info-photo > img {
-  width: 250px;
-
-  border-radius: 100%;
-  background: linear-gradient(
-    to top,
-    var(--primary-color),
-    var(--primary-color-light)
-  );
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.19);
-}
-
-.profile-container--info-me > p {
-  width: 400px;
-}
-
-.profile-container--info-me > p:nth-child(1) {
-  background: var(--gradient-left-right);
-  -webkit-background-clip: text;
-  background-clip: text;
-  color: transparent;
-}
-
-.profile-container--info-me > p:nth-child(1) {
-  font-size: 45px;
-  font-weight: bold;
-  line-height: 1;
-  margin-bottom: 30px;
-}
-
-.profile-container--info-me > p:nth-child(2) {
-  font-size: 20px;
-  font-weight: 500;
-  line-height: 1;
-}
-
-.profile-container--info-me > p:nth-child(3) {
-  font-size: 15px;
-
-  line-height: 1.5;
-}
-.profile-container--info-me > p > span {
-  color: var(--primary-color);
-}
-
-@media (max-width: 752px) {
-  .profile-container--info-photo > img {
-    width: 50%;
-    min-width: 150px;
-    margin-left: 0;
-    padding: 0px;
-  }
-
-  .profile-container--info {
-    justify-content: center;
-    align-items: center;
-    flex-direction: column-reverse;
-    text-align: center;
-  }
-
-  .profile-container--info-me > p {
-    width: 100%;
-    max-width: 400px;
-  }
-
-  .profile-container--info-me > p:nth-child(1) {
-    font-size: 2em;
-  }
-
-  .profile-container--info-me > p:nth-child(2) {
-    font-size: 1em;
-  }
-
-  .profile-container--info-me > p:nth-child(3) {
-    font-size: 1em;
-  }
-}
-@media (max-width: 400px) {
-  .profile-container--info-me > p {
-    width: 250px;
-  }
-}
