+++ src/theme/colors.scss
@@ -1,0 +1,2 @@
+$primary-color: #127bb3;
+$primary-color-light: #00a0f7;
