+++ src/theme/styles.scss
@@ -1,0 +1,6 @@
+@use './colors.scss' as *;
+
+$box-shadow-card: 0 4px 6px rgba(0, 0, 0, 0.1);
+$box-shadow-card-medium: 0 6px 12px rgba(0, 0, 0, 0.2);
+$gradient-left-right: linear-gradient(to right, black, $primary-color-light);
+$gradient-bottom-topRight: linear-gradient(to left bottom, $primary-color, $primary-color-light);
